Remove CLA, add note about bots
The CLA wording was unnecessary. However, I think it is important to let users know that they may run into bots which enforce similar aspects of interacting with the repository; for instance, stale bot, CLA bots, and so on. Adding a note in the Contributing about various bots that work on the platform would be a good way of doing this that I have not seen widely adopted (read: at all) yet. But I think it would help here.